backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["jordanhoare <jordanhoare@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
httpx = "^0.23.3"
cassandra-driver = "^3.25.0"
pydantic = {extras = ["dotenv"], version = "^1.10.4"}
pandas = "^1.5.3"
faker-vehicle = "^0.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
pytest-cov = "^3.0.0"
black = "^22.6.0"
mypy = "^0.971"
isort = "^5.10.1"
flake8 = "^5.0.4"
poethepoet = "^0.18.1"


[tool.black]
line-length = 99
target-version = ["py38"]

[tool.isort]
profile = 'black'

[tool.poe.tasks]
run = "poetry run python -m app" 
clean = """
poetry run black . &
poetry run isort . --profile black &
poetry run mypy . &
poetry run flake8 app
""" 

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

